ml-service/requirements.txt
# FastAPI - Web framework for ML service
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning Libraries
prophet==1.1.5
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Database
pymongo==4.6.0
motor==3.3.2

# Environment variables
python-dotenv==1.0.0

# Google Gemini API (Free LLM)
google-generativeai==0.3.1

# Utilities
joblib==1.3.2
pydantic==2.5.0
python-multipart==0.0.6

# CORS
fastapi-cors==0.0.6
